requirements.txt
matplotlib==3.1.1
numpy==1.17.0
Pillow==6.2.0
python-dotenv==0.10.3
requests==2.22.0
scipy==1.3.1
tensorflow==2.0.0